新增Event事件功能
新增内容:
	X_Skyworth.Upload()
	X_Skyworth.UpgradeFile()
	X_Skyworth.DownloadFile()
	Device.IP.Diagnostics.IPPing()
	Device.IP.Diagnostics.TraceRoute()
	Device.IP.Diagnostics.DownloadDiagnostics()

+++ accessor/src/main/java/com/sdt/diagnose/Device/X_Skyworth/Tcpdump/TcpdumpBean.java
@@ -12,13 +12,13 @@ public class TcpdumpBean {
     String netType;
     String duration;
     String url;
-    int fileSize;
+    long fileSize;
 
-    public int getFileSize() {
+    public long getFileSize() {
         return fileSize;
     }
 
-    public void setFileSize(int fileSize) {
+    public void setFileSize(long fileSize) {
         this.fileSize = fileSize;
     }
